+++ satella/coding/concurrent/queue.py
@@ -75,18 +75,21 @@ class PeekableQueue(tp.Generic[T]):
 
     @rethrow_as(WouldWaitMore, Empty)
     def __get(self, timeout, item_getter) -> T:
-        self.lock.acquire()
-        if len(self.queue):
-            # Fast path
-            try:
-                return item_getter(self.queue)
-            finally:
-                self.lock.release()
-        else:
-            if timeout is None:
-                return self.__get_timeout_none(item_getter)
+        try:
+            self.lock.acquire()
+            if len(self.queue):
+                # Fast path
+                try:
+                    return item_getter(self.queue)
+                finally:
+                    self.lock.release()
             else:
-                return self.__get_timeout(item_getter, timeout)
+                if timeout is None:
+                    return self.__get_timeout_none(item_getter)
+                else:
+                    return self.__get_timeout(item_getter, timeout)
+        finally:
+            self.items_count -= 1
 
     def get(self, timeout: tp.Optional[float] = None) -> T:
         """
@@ -97,7 +100,10 @@ class PeekableQueue(tp.Generic[T]):
         :return: the item
         :raise Empty: queue was empty
         """
-        return self.__get(timeout, lambda queue: queue.popleft())
+        try:
+            return self.__get(timeout, lambda queue: queue.popleft())
+        finally:
+            self.items_count -= 1
 
     def peek(self, timeout: tp.Optional[float] = None) -> T:
         """