+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v.25.6.
+# v.25.a3
 
 * fixed circular import